# Speed up ILEmitTrie jump table using Binary Search Tree