@adityamandaleeka 2024-01-10
Thanks for the effort here @andrewjsaid. This looks well thought-out and I appreciate the perf numbers you collected. Tagging @javiercn and @mkArtakMSFT for review since they own the routing area.